// src/app/game/state/game-socket-actions.enum.ts
export enum GameSocketActions {
  GAMING_OPPONENT_FOUND = 'GAMING_OPPONENT_FOUND',
  GAMING_SUBSCRIBE_TO_BOARD = 'GAMING_SUBSCRIBE_TO_BOARD',
  GAMING_GAME_STARTED = 'GAMING_GAME_STARTED',
  GAMING_ADD_MOVE = 'GAMING_ADD_MOVE',
  GAMING_NOTIFICATION = 'GAMING_NOTIFICATION',
  GAMING_BOARD_CREATED = 'GAMING_BOARD_CREATED',
  GAMING_GAME_TIME_CONTROL_END = 'GAMING_GAME_TIME_CONTROL_END',
  GAMING_GAME_END = 'GAMING_GAME_END',
  GAMING_GAMING_RESIGN = 'GAMING_RESIGN',
  GAMING_GAME_TIMEOUT_END = 'GAMING_GAME_TIMEOUT_END',
  GAMING_DRAW_OFFER = 'GAMING_DRAW_OFFER',
  GAMING_RATING_CHANGE = 'GAMING_RATING_CHANGE',
  GAMING_GAME_ABORT = 'GAMING_GAME_ABORT',
  GAMING_PGN_CREATED = 'GAMING_PGN_CREATED',
  GAMING_PING = 'GAMING_PING',
  GAMING_PONG = 'GAMING_PONG',
  GAME_DISCONNECT = 'GAME_DISCONNECT',
}